# Optimize parser by a Zillion %