The main problem was repeated allocations when consuming characters by doing

```Crystal
@buffer += @str[@i]
```

So to parse the string `abc` we allocated `a`, then `ab`, then `abc`.

Now the code just store the string indexes and use them when necessary. So a 400kb file that was taking more than a hour to render now render in milliseconds.

Tested with a plain text file with 200kb then with a 2Mb file got the results:

```
$ ./parser foo/sample3.txt
rendering foo/sample3.txt
  0.005337   0.000000   0.005337 (  0.005346)
$ ./parser foo/sample-2mb-text-file.txt
rendering foo/sample-2mb-text-file.txt
  0.057601   0.000000   0.057601 (  0.057687)
$ ls -lah foo/
Permissions Size User Date Modified Name
.rw-r--r--   202 hugo 25 nov 14:57  parser.cr
.rw-r--r--  2,2M hugo 25 nov 14:55  sample-2mb-text-file.txt
.rw-r--r--  177k hugo 25 nov 14:55  sample3.txt
```

Test env:
 - ArchLinux
 - Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
 - 16Gb RAM
 - Some Sata SSD hard drive
 - Crystal 1.6.2
 - LLVM: 14.0.6