@robacarp 2022-11-29
This is quite the optimization. I have two questions.

First, this change makes the code far less obvious to the reader that's not familiar with the code. The consequence of this is that later bugs are less likely to be found and fixed by any arbitrary community member. Perhaps that's the problem domain at play. However, I wonder if it would be possible to use a String::Builder class instead? I think it uses a similar algorithm behind the scenes. Would that make it easier to read?

Second, with any performance related pull request, I would like to see a before and after comparison. Ideally we would have performance specs built into CI so an easy comparison could be made. In lieu of that, would you please provide a before and after measurement? Crystal provides a nice [Benchmark](https://crystal-lang.org/api/1.2.2/Benchmark.html) class which runs the simulation multiple times and performs the statistics automatically.