@hugopl 2022-11-30
Hi,

Yes, more optimized code are most of the time a bit harder for the eyes, that's the price.

BTW I guess I did something wrong with the first results (I guess I forgot to recompile the benchmark and measured the speed of a lexer I was doing as experiment :sweat_smile: )

But here the results of parsing the 345kb file of `Int8.html` file from Crystal docs.

**Current master branch**
```
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  6.520321   0.026642   6.546963 (  6.550224)
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  6.015328   0.012993   6.028321 (  6.031401)
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  5.987136   0.016721   6.003857 (  6.007143)
```

**This PR**
```
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  0.004232   0.000000   0.004232 (  0.004239)
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  0.004174   0.000000   0.004174 (  0.004179)
$ ./parser-test /usr/share/doc/crystal/api/Int8.html 
  0.004277   0.000000   0.004277 (  0.004282)
```

**Benchmark code**
```Crystal
require "benchmark"
require "./src/liquid"

filename = ARGV[0]
contents = File.read(filename)

time = Benchmark.measure do
  Liquid::Template.parse(contents)
end

 puts time
```

The file used as template in the benchmark has no liquid tags at all, in production I need to render 400kb files and the Crystal version of liquid still *a lot* slower than the Ruby version even with this optimization... my guess is the over used of `String#[]` that's causing the slowness, I have a branch that reads the string using `Char::Reader`, however I noticed that the Ruby version of liquid does the entire tokenization just using regexes (what explain liquid not support things like `{{ "}}" }}`), maybe this is the way to go... but it would require I rewrite of the Parser class.

**Benchmark Env**
ArchLinux, Crystal 1.6.2, LLVM 14.0.6, AMD Ryzen 9 5900X 12-Core Processor, 32G RAM