@hugopl 2022-11-30
> I wonder if it would be possible to use a String::Builder class instead? I think it uses a similar algorithm behind the scenes. Would that make it easier to read?

`String::Builder` allocates an initial buffer then calls `realloc` when the buffer is full and more data is written, so it would be faster than the current code in master due to the larges initial and reallocated buffers, however slower than this PR approach that does no `realloc` calls at all.